refactor: split monolithic test file into 8 focused test files
Split the 4,812-line src/ralphai.test.ts into 8 domain-specific test files
to improve maintainability and keep files under the 500-line test file limit.

New files: init, uninstall-reset, package-manager, cli-help, worktree,
runner-config, self-update, status-doctor (264 tests total, unchanged).

Also adds file size guardrails to AGENTS.md and a shared useTempGitDir()
helper to test-utils.ts.
